% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_track}
\alias{fit_track}
\title{Fit a Gaussian process model to an sf data.frame of location pings.}
\usage{
fit_track(
  pings,
  track_time,
  time_units = "days",
  covariance_code = 3,
  number_of_ping_parents = 5,
  number_of_track_parents = 5,
  track_robustness = 6,
  track_robust_code = 0,
  ping_robustness = 6,
  ping_robust_code = 2,
  starting_move_sd = c(12, 12),
  starting_move_range = c(3, 3),
  starting_ping_sd = c(500, 500),
  starting_ping_cor = 0,
  fix_move_sd = FALSE,
  fix_move_range = TRUE,
  fix_ping_sd = FALSE,
  fix_ping_cor = TRUE,
  k = 5,
  silent = TRUE,
  ...
)
}
\arguments{
\item{pings}{An sf data.frame with  point geometries, a "time" column, and a "quality_class" column}

\item{track_time}{Optional vector of times to use for modeling the true locations. Defaults to observed ping times.}

\item{time_units}{Units to use when computing time differences}

\item{covariance_code}{0 = exponential, 1 = gaussian, 2 = matern, 3 = matern32}

\item{number_of_ping_parents}{Number of observation pings to use as parents}

\item{number_of_track_parents}{Number of directly prior track locations to use as parents}

\item{track_robustness}{Tuning parameters for robustifying process}

\item{track_robust_code}{Robustification type: 0 = ML, 1 = arctan, 2 = smooth semi-Huber}

\item{ping_robustness}{Tuning parameters for robustifying observations}

\item{ping_robust_code}{Robustification type: 0 = ML, 1 = arctan, 2 = smooth semi-Huber}

\item{starting_move_sd}{Starting values for std. dev. of movement velocity components}

\item{starting_move_range}{Starting values for movement range parameter}

\item{starting_ping_sd}{Starting values for observation error std. dev. (ARGOS class 3) (measured in m)}

\item{starting_ping_cor}{Starting values for observation error ellipse correlation}

\item{fix_move_sd}{Should these parameters be held fixed in the estimation?}

\item{fix_move_range}{Should these parameters be held fixed in the estimation?}

\item{fix_ping_sd}{Should these parameters be held fixed in the estimation?}

\item{fix_ping_cor}{Should these parameters be held fixed in the estimation?}

\item{k}{Length of sliding window for median filter on location estimates.}

\item{silent}{Should printing of tracing information be prevented}

\item{...}{Options passed to nlminb}
}
\value{
A list with pings, estimated track, TMB object, optimization object, and sdreport object.
}
\description{
Fit a Gaussian process model to an sf data.frame of location pings.
}
