% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loc_class_K}
\alias{loc_class_K}
\title{Location quality class error multipliers}
\format{
An object of class \code{data.frame} with 7 rows and 3 columns.
}
\source{
Taken from the foieGras R package.
}
\usage{
loc_class_K
}
\description{
Location quality class error multipliers
}
\keyword{datasets}
