% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_posterior_track_profiles}
\alias{plot_posterior_track_profiles}
\title{Plot the coordinate profiles of a track}
\usage{
plot_posterior_track_profiles(
  track,
  posterior_tracks,
  ping_weights,
  lower_weight_limit = 0.01,
  alpha = 0.2,
  point_size_scale = 1,
  which_samples = seq_along(posterior_tracks)
)
}
\arguments{
\item{track}{A list with a pings sf data.frame and a track sf data.frame.}

\item{posterior_tracks}{The output of sample_posterior_tracks()}

\item{ping_weights}{An optional vector of score-level weights for each ping}

\item{lower_weight_limit}{Pings with weight below this limit will not be plotted.}

\item{alpha}{Transparency value for posterior samples. 0 is invisible, 1 is opaque.}

\item{point_size_scale}{Scaling factor for point sizes}

\item{which_sample}{Index vector of which samples to plot}
}
\value{
A tmap object
}
\description{
Plot the coordinate profiles of a track
}
