% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_track}
\alias{plot_track}
\title{Plot a track}
\usage{
plot_track(
  track,
  ping_weights,
  lower_weight_limit = 0.01,
  point_size_scale = 1,
  return_type = "plot"
)
}
\arguments{
\item{track}{A list with a pings sf data.frame and a track sf data.frame.}

\item{ping_weights}{An optional vector of score-level weights for each ping}

\item{lower_weight_limit}{Pings with weight below this limit will not be plotted.}

\item{point_size_scale}{Scaling factor for point sizes}

\item{return_type}{Either "plot" to return a plot, or "list" to return a list of plot elements.}
}
\value{
A tmap object
}
\description{
Plot a track
}
