% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_track_profile}
\alias{plot_track_profile}
\title{Plot the coordinate profiles of a track}
\usage{
plot_track_profile(
  track,
  x_se,
  y_se,
  ping_weights,
  lower_weight_limit,
  point_size_scale = 1
)
}
\arguments{
\item{track}{A list with a pings sf data.frame and a track sf data.frame.}

\item{x_se, }{y_se Optional vector of standard errors for the x and y coordinate.}

\item{ping_weights}{An optional vector of score-level weights for each ping}

\item{lower_weight_limit}{Pings with weight below this limit will not be plotted.}

\item{point_size_scale}{Scaling factor for point sizes}
}
\value{
A tmap object
}
\description{
Plot the coordinate profiles of a track
}
