% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{predict_track}
\alias{predict_track}
\title{Interpolate a fitted track at new dates.}
\usage{
predict_track(fit, date, quantiles = c(0.05, 0.5, 0.95))
}
\arguments{
\item{fit}{A fitted model from fit_track.}

\item{date}{The dates at which to predict.}

\item{quantiles}{The quantiles to predict. Defaults to 0.05, 0.5, and 0.95.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A list with the predicted quantiles of the coordinates at each date and
the posterior track samples. If length(quantiles) < 1 then only the
posterior samples are returned.
}\if{html}{\out{</div>}}
}
\description{
Interpolate a fitted track at new dates.
}
