% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_posterior.R
\name{sample_posterior_tracks}
\alias{sample_posterior_tracks}
\title{Sample a track from the posterior of a fitted model}
\usage{
sample_posterior_tracks(
  track_fit,
  number_of_samples = 10,
  method = "joint_precision",
  n_cores = 1
)
}
\arguments{
\item{track_fit}{The output of fit_track()}

\item{number_of_samples}{The number of posterior samples}

\item{method}{Either "joint_precision" or "posterior_nngp". If "joint_precision":
posterior samples are obtained by inverting the joint precision matrix
and sampling directly from the multivariate normal distribution with mean
equal to the point estimates.

If "posterior_nngp": correlation parameters are sampled from the point estimates
and fixed effects precision matrix. The marginal mean and standard deviation
of each random effect is set to the point estimate and standard error from
the model fit. Then the above quantities are fed into the NNGP model for simulation.}

\item{n_cores}{Number of parallel cores to use.}
}
\value{
A list containing the posterior sampled tracks
}
\description{
Sample a track from the posterior of a fitted model
}
