% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_track}
\alias{simulate_track}
\title{Simulate a track with pings}
\usage{
simulate_track(
  track_time = seq(0, 1, length.out = 101),
  ping_time = seq(0, 1, length.out = 25),
  endpoint = rbind(c(0, 0), c(0, 0)),
  covariance_code = 1,
  covariance_parameters = rbind(c(1, 0.2, Inf), c(1, 0.2, Inf)),
  ping_parameters = c(0.1, 0, 0.1),
  quality_class_probabilities = c(0.1, 0.1, 0.2, 0.4, 0.2, 0, 0),
  number_of_ping_parents = 5,
  number_of_track_parents = 5,
  seed
)
}
\arguments{
\item{track_time}{A vector giving the timestamps of track locations}

\item{ping_time}{A vector giving the timestamps of location pings}

\item{endpoint}{The (expected) starting and ending points}

\item{covariance_code}{0 = exponential, 1 = gaussian, 2 = matern, 3 = matern32}

\item{covariance_parameters}{A 2x3 matrix. Columns: sd, range, smoothness}

\item{ping_parameters}{A 3-vector giving the reference observation error}

\item{quality_class_probabilities}{A vector giving relative frequencies of quality classes}

\item{number_of_ping_parents}{Number of observation pings to use as parents}

\item{number_of_track_parents}{Number of directly prior track locations to use as parents}

\item{seed}{Random number seed}
}
\value{
A list
}
\description{
Simulate a track with pings
}
