% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{split_track}
\alias{split_track}
\title{Split a track into subsets}
\usage{
split_track(pings, track_time, split_percentage = 1, split_overlap = 0)
}
\arguments{
\item{pings}{An sf data.frame with point geometries, a "time" column, and a "quality class" column}

\item{track_time}{Optional vector of times to use for modeling the true locations. Defaults to observed ping times.}

\item{split_percentage}{What percentage of the observed data should each split contain?}

\item{split_overlap}{How much should each split overlap adjacent splits? Each split will be extended to cover the overlap.}
}
\value{
A list giving the splits of the pings and track times.
}
\description{
Split a track into subsets
}
