% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackfiltering-package.R
\docType{package}
\name{trackfiltering-package}
\alias{trackfiltering}
\alias{trackfiltering-package}
\title{trackfiltering: Robust filtering for movement path estimation}
\description{
Estimates a true movement path from a noisy set of telemetry data, automatically detecting and removing the effect of outliers.
}
\author{
\strong{Maintainer}: Ethan Lawler \email{lawlerem@gmail.com} (\href{https://orcid.org/0000-0002-6225-3009}{ORCID})

}
\keyword{internal}
