% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_track}
\alias{fit_track}
\title{Fit a model to a movement track}
\usage{
fit_track(
  pings,
  time_mesh = unique(pings$date),
  time_units = "days",
  independent_coordinates = FALSE,
  common_coordinate_correlation = TRUE,
  ...
)
}
\arguments{
\item{pings}{An sf data frame with a POSIX or numeric column named "date",
a factor column named "class", and point geometries. If class is an
ordered factor, then the estimated observation error variances will
increase with class.}

\item{time_mesh}{A vector of times used to discretize movement.}

\item{time_units}{Which time units should be used? Mostly matters for numerical stability.}

\item{independent_coordinates}{If true, then observation ellipses are assumed to have independent axes.}

\item{common_coordinate_correlation}{If true, then observation ellipses for different quality classes are
assumed to have the same correlation structure.}

\item{...}{Additional arguments, particularly those to pass to
nnspline::create_lcspline, robustifyRTMB::robustly_optimize,
and mcreportRTMB::mcreport.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A list with the following:
}\if{html}{\out{</div>}}
\itemize{
\item pings
A copy of the provided location pings with an additional weights
column giving the robust weight of each observation, an "est"
column(s) giving the estimated locations, and an "se" column(s)
giving the standard errors around the estimated locations.
\item error_matrices
A list of estimated observation error covariance matrices for each
quality class.
\item template_spline
An nnspline object used to hold the nodes and node graph used for
the model.
\item start_time
The date used for t = 0.
\item robust_optimization
The output of robustifyRTMB::robustly_optimize
\item mcreport
The output of mcreportRTMB::mcreport for the spline parameters,
node values, and coordinate means.
\item call
The function call used.
}
}
\description{
Fit a model to a movement track
}
