% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_track}
\alias{fit_track}
\title{Fit a model to a movement track}
\usage{
fit_track(
  pings,
  interpolation_time = numeric(0),
  nodes = time,
  robust_schedule = seq(0, 0.05, by = 0.01),
  robust_function = "ll",
  robust_bandwidth = 0.001 * diff(range(time)),
  independent_coordinates = FALSE,
  common_coordinate_correlation = TRUE,
  time_units = "days"
)
}
\arguments{
\item{pings}{An sf data frame with a POSIX or numeric column named "date", an optional factor column named "class", and point geometries.}

\item{nodes}{A vector of times used as nodes in the underlying spline.}

\item{robust_schedule}{See robustifyRTMB::robustly_optimize}

\item{robust_function}{See robustifyRTMB::robustify}

\item{robust_bandwidth}{See robustifyRTMB::robustly_optimize}

\item{independent_coordinates}{If true, then observation ellipses are assumed to have
independent axes.}

\item{common_coordinate_correlation}{If true, then observation ellipses for different
quality classes are assumed to have the same correlation structure.}

\item{time_units}{Which time units should be used? Mostly matters for numerical stability.}

\item{interpolate_time}{A vector of times at which to interpolate locations.}
}
\description{
Fit a model to a movement track
}
